requirements.txt
Pillow==10.3.0
Requests==2.31.0
streamlit==1.33.0
streamlit_lottie==0.0.5
